requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.9
pillow==11.1.0
paddlepaddle==2.6.2
paddleocr==2.7.3
opencv-python-headless==4.9.0.80
numpy<2.0.0

